Moved methods
- validate and prune moved to GameManager class to abstract from
Gameboard
- removed consumeInput() from Gameboard and added play() that takes its
place in GameManager
- removed GameboardResponse and moved its function to GameManager
